Allow the gitignore filter to work on name components only
We end up with a lot of rules like (?:.*)/.*\.ext which are basically
trying to find the last component in a path and match against
that. These are rather slow to run so the easiest thing tdo is just
pass in the last component of the path when we know that's the only
thing the rule can match.

The changes to surrounding code to use this API will be made in future
commits.

Depends on D8222

Differential Revision: https://phabricator.services.mozilla.com/D8223

bugzilla-url: https://bugzilla.mozilla.org/show_bug.cgi?id=1497898
gecko-commit: 4082767a2fbf8ab4283e4b39e72bddfc6d4bb198
gecko-integration-branch: central
gecko-reviewers: ato