Fix the cache lifecycle
The caches weren't being invalidated when the manifest itself
changed. To fix this the manifest itself has to be written before the
cache and the cache has to include data about the manifest that it's
associated with (the mtime and path are used for this purpose).

To make all this work requires a single method that can load the
manifest, update it, write the manifest and write the
caches. Therefore we introduce a single load_and_update method that is
intended to replace all previous use of the load() or update() methods
(and as a bonus handles manifest version mismatches in a single
place).

Depends on D8227

Differential Revision: https://phabricator.services.mozilla.com/D8228

bugzilla-url: https://bugzilla.mozilla.org/show_bug.cgi?id=1497898
gecko-commit: cd31e38cc943c12c6bce54d66b67c93b2793ee05
gecko-integration-branch: central
gecko-reviewers: ato